Refactor to better separate concerns. Added tests
