+++ src/config/mod.rs
@@ -1,0 +1,2 @@
+pub const AUTH_SECRET: &str = "your_secret_key";
+pub const JWT_EXPIRATION_SECS: u64 = 3600;
